requirements.txt
schedule
pyTelegramBotAPI
requests
emoji
python-dotenv
gigachat
